+++ src/Exceptions/NotSupportedException.php
@@ -5,6 +5,8 @@ namespace Saritasa\Transformers\Exceptions;
 /**
  * Exception class is thrown when trying to use features not supported by classes
  * or abstractions.
+ *
+ * @deprecated Use \Saritasa\NotImplementedException instead
  */
 class NotSupportedException extends \RuntimeException
 {